requirements.txt
typer[all]
mysql-connector-python
pytest
pytest-docker
ruff
jinja2
pystardog
